implementing select2 + handling deleting tags and categories associated toexisting posts + handling user profiles + turning regular users to admin users feature+ updating user profiles feature + installing a frontend theme + making changes to welcome page

+++ app/Http/Controllers/CategoriesController.php
@@ -97,6 +97,12 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if($category->posts->count() > 0){
+            session()->flash('error','Category Cannot be deleted because it is associated to some posts.');
+
+            return redirect()->back();
+        }
+        
         $category->delete();
 
         session()->flash('success','Category Deleted Successfully.');
