+++ app/Http/Controllers/TagsController.php
@@ -97,6 +97,12 @@ class tagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if($tag->posts->count() > 0){
+            session()->flash('error','Tag Cannot be deleted because it is associated to some posts.');
+
+            return redirect()->back();
+        }
+        
         $tag->delete();
 
         session()->flash('success','Tag Deleted Successfully.');
